==> scripts/03_operational_processing.ipynb <==
# 1. IMPORTS

from datetime import timedelta
import pandas as pd
import os

# 2. PATHS

BASE_PATH = os.path.dirname(os.getcwd())
OPERATIONAL_PATH = os.path.join(BASE_PATH, "data", "operational")
PROCESSED_PATH = os.path.join(BASE_PATH, "data", "processed")
os.makedirs(PROCESSED_PATH, exist_ok=True)

# 3. LOAD DATAFRAME

cord_interventions = pd.read_excel(
    os.path.join(OPERATIONAL_PATH, "cord_interventions.xlsx"),
    header=2,
    dtype={'Student_ID': str}  # garante que o ID seja string
)

display(cord_interventions)

# 4. DATA TYPE TREATMENT

cord_interventions['Intervention_Date'] = pd.to_datetime(
    cord_interventions['Intervention_Date'], errors='coerce'
)

# 5. FILTER LAST 30 DAYS

today = pd.Timestamp("2026-02-12")

# Filtra apenas os últimos 30 dias
recent_interventions = cord_interventions.loc[
    cord_interventions['Intervention_Date'].between(today - timedelta(days=30), today),
    ['Student_ID', 'Intervention_Type', 'Intervention_Description', 'Case_Status', 'Intervention_Date']
]

display(recent_interventions)

# SAVE PROCESSED DATAFRAME

recent_interventions.to_csv(
    os.path.join(PROCESSED_PATH, "recent_interventions.csv"),
    index=False,
    encoding="utf-8-sig"
)

print("Processed data exported successfully.")
